Fixed Code Review
- e는 UI 컴포넌트에서 관리
- 함수명 변경
