+++ src/App.jsx
@@ -6,34 +6,23 @@ import TodoTemplate from './components/TodoTemplate';
 
 function App() {
   const [todos, setTodos] = useState([]);
-  const [todo, setTodo] = useState('');
-
-  const handleOnChange = (e) => {
-    setTodo(e.target.value);
-  };
 
   const onInsert = (value) => {
-    const item = {
+    const todo = {
       id: todos.length ? todos[todos.length - 1].id + 1 : 1,
       value,
     };
 
-    setTodos([...todos, item]);
-  };
-
-  const onSubmit = (e) => {
-    e.preventDefault();
-    onInsert(todo);
-    setTodo('');
+    setTodos([...todos, todo]);
   };
 
   const handleRemoveTodo = (id) => {
-    setTodos(todos.filter((item) => item.id !== id));
+    setTodos(todos.filter((todo) => todo.id !== id));
   };
 
   return (
     <TodoTemplate>
-      <TodoInsert onChange={handleOnChange} onSubmit={onSubmit} todo={todo} />
+      <TodoInsert insert={onInsert} />
       <TodoList todos={todos} onRemove={handleRemoveTodo} />
     </TodoTemplate>
   );
